+++ src/main/java/com/braintreegateway/VisaCheckoutCard.java
@@ -7,6 +7,8 @@ import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
 
+// NEXT_MAJOR_VERSION remove isVenmoSDK
+// The old venmo SDK integration has been deprecated
 public class VisaCheckoutCard implements PaymentMethod {
 
     private Address billingAddress;
@@ -20,6 +22,7 @@ public class VisaCheckoutCard implements PaymentMethod {
     private String expirationMonth;
     private String expirationYear;
     private boolean isDefault;
+    @Deprecated
     private boolean isVenmoSdk;
     private boolean isExpired;
     private String imageUrl;
@@ -255,6 +258,10 @@ public class VisaCheckoutCard implements PaymentMethod {
         return isDefault;
     }
 
+    //NEXT_MAJOR_VERSION remove this method
+    /**
+     * @deprecated - The Venmo SDK integration is Unsupported. Please update your integration to use Pay with Venmo instead
+    */
     public boolean isVenmoSdk() {
       return isVenmoSdk;
     }